Refactor KMeans benchmarks; make Scala and Java implementations work independently.
