--- KMeans/src/main/java/SimpleApp.java
@@ -1,43 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
-package kmeans_java;
-import org.apache.log4j.Logger;
-import org.apache.log4j.Level;
-import org.apache.spark.api.java.*;
-import org.apache.spark.SparkConf;
-import org.apache.spark.api.java.function.Function;
-/**
- *
- * @author minli
- */
-public class SimpleApp {
-
-    public static void main(String[] args) {
-Logger.getLogger("org.eclipse.jetty.server").setLevel(Level.OFF);
-Logger.getLogger("org.apache.spark").setLevel(Level.WARN);
-
-        String logFile = "hdfs:///README.md"; // Should be some file on your system
-        SparkConf conf = new SparkConf().setAppName("Simple Application");
-        JavaSparkContext sc = new JavaSparkContext(conf);
-        JavaRDD<String> logData = sc.textFile(logFile).cache();
-
-        long numAs = logData.filter(new Function<String, Boolean>() {
-            public Boolean call(String s) {
-                return s.contains("a");
-            }
-        }).count();
-
-        long numBs = logData.filter(new Function<String, Boolean>() {
-            public Boolean call(String s) {
-                return s.contains("b");
-            }
-        }).count();
-
-        System.out.println("Lines with a: " + numAs + ", lines with b: " + numBs);
-        sc.stop();
-    }
-}
