+++ KMeans/src/main/scala/kmeans.scala
@@ -1,4 +1,3 @@
-
 /*
  * (C) Copyright IBM Corp. 2015 
  *
@@ -15,21 +14,17 @@
  * limitations under the License.
  */
 
-import org.apache.log4j.Logger
-import org.apache.log4j.Level
-import org.apache.spark.mllib.clustering.{KMeansModel, KMeans}
-import org.apache.spark.{SparkContext,SparkConf}
-import org.apache.spark.SparkContext._
+import org.apache.log4j.{Level, Logger}
+import org.apache.spark.mllib.clustering.{KMeans, KMeansModel}
 import org.apache.spark.mllib.linalg.Vectors
-
-import org.json4s.JsonAST._
+import org.apache.spark.{SparkConf, SparkContext}
 import org.json4s.JsonDSL._
 import org.json4s.jackson.JsonMethods._
 
 object KmeansApp {
   def main(args: Array[String]) {
-Logger.getLogger("org.apache.spark").setLevel(Level.WARN);
-Logger.getLogger("org.eclipse.jetty.server").setLevel(Level.OFF);
+    Logger.getLogger("org.apache.spark").setLevel(Level.WARN);
+    Logger.getLogger("org.eclipse.jetty.server").setLevel(Level.OFF);
     if (args.length < 4) {
       println("usage: <input> <output> <numClusters> <maxIterations> <runs> - optional")
       System.exit(0)
@@ -37,7 +32,7 @@ Logger.getLogger("org.eclipse.jetty.server").setLevel(Level.OFF);
     val conf = new SparkConf
     conf.setAppName("Spark KMeans Example")
     val sc = new SparkContext(conf)
-    
+
     val input = args(0)
     val output = args(1)
     val K = args(2).toInt
@@ -53,12 +48,12 @@ Logger.getLogger("org.eclipse.jetty.server").setLevel(Level.OFF);
 
     // Cluster the data into two classes using KMeans
     start = System.currentTimeMillis();
-    val clusters: KMeansModel = KMeans.train(parsedData, K, maxIterations,runs, KMeans.K_MEANS_PARALLEL, seed=127L)
+    val clusters: KMeansModel = KMeans.train(parsedData, K, maxIterations, runs, KMeans.K_MEANS_PARALLEL, seed = 127L)
     val trainingTime = (System.currentTimeMillis() - start).toDouble / 1000.0
     println("cluster centers: " + clusters.clusterCenters.mkString(","))
 
     start = System.currentTimeMillis();
-    val vectorsAndClusterIdx = parsedData.map{ point =>
+    val vectorsAndClusterIdx = parsedData.map { point =>
       val prediction = clusters.predict(point)
       (point.toString, prediction)
     }
@@ -74,6 +69,7 @@ Logger.getLogger("org.eclipse.jetty.server").setLevel(Level.OFF);
     println("Within Set Sum of Squared Errors = " + WSSSE)
     sc.stop()
   }
+
   def calculateRuns(args: Array[String]): Int = {
     if (args.length > 4) args(4).toInt
     else 1
